requirements.txt
# Requirements for Thesis Tests Microservice
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.10

# Configuration
pydantic-settings==2.6.1
python-dotenv==1.0.1

# PDF Processing
pdfplumber==0.11.4
pdf2image==1.17.0
Pillow==11.0.0
camelot-py[base]>=1.0.9

# File uploads
python-multipart==0.0.20

# Cache
redis==5.2.1
